Removed extraneous files.
When moving files to a new directory in git, you have to
add the directory to move them AS WELL AS add the names
of the moved files to delete them from the previous directory!
I just did that because I did not realize that I had to

Also writing long git commit messages should be done without
the -m option in the git commit command. Instead edit the file
that is opened in vim with the normal git commit command
(like I am doing now) and put the super brief description in
the first line, then skip a line, and write the rest (being
sure to manually wrap text at desired line length, like I am
doing now.)